move to components.py
* later will merge template_mixin into here as well
